requirements.txt
django>=4.1,<4.2
djangorestframework
pillow
psycopg2-binary
requests
django-filter